updated navbar and enterpage
