+++ src/Components/CardRender/CardRender.jsx
@@ -15,8 +15,6 @@ import "react-toastify/dist/ReactToastify.css";
 const serverAddress = import.meta.env.VITE_DEV_URL
 
 
-
-
 pokemon.configure({ apiKey: import.meta.env.VITE_TCG_API });
 
 export default function CardRender() {
@@ -25,7 +23,6 @@ export default function CardRender() {
   const [css, setCss] = useState(false);
   //post request states
   const [userName, setUserName] = useState("");
-
   const [bookletData, setBookletData] = useState([]);
 
   useEffect(() => {
