+++ src/Pages/App/App.css
@@ -1,9 +1,9 @@
-#root {
+/* #root {
   max-width: 1280px;
   margin: 0 auto;
   padding: 2rem;
   text-align: center;
-}
+} */
 
 img {
   color: black;
@@ -86,6 +86,20 @@ img {
   height: 20.625rem;
   caret-color: transparent;
   margin: 4rem auto;
+}
+
+.iphone__screen {
+  max-width: 1280px;
+  margin: 0 auto;
+  padding: 2rem;
+  text-align: center;
+}
+
+.enterPage {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 @media screen and (max-width: 576px) {
