+++ src/Pages/App/App.jsx
@@ -6,9 +6,8 @@ import { HTML5toTouch } from 'rdndmb-html5-to-touch' // or any other pipeline
 import {Route,Routes} from 'react-router-dom';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import CardRender from "../../Components/CardRender/CardRender";
-import EnterPage from "../../Components/EnterPage/EnterPage";
-
+import EnterPage from "../EnterPage/EnterPage";
+import Home from "../Home/Home";
 
 
 function App() {
@@ -18,7 +17,7 @@ function App() {
         position = {toast.POSITION.BOTTOM_CENTER}/>
         <Routes>
           <Route path="/" element={<EnterPage />} />
-          <Route path="/home" element={<CardRender />} />
+          <Route path="/home" element={<Home />} />
         </Routes>
       </DndProvider>
   );
