+++ src/Pages/Home/Home.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import NavBar from '../../Components/NavBar/NavBar'
+import CardRender from '../../Components/CardRender/CardRender'
+
+export default function Home() {
+  return (
+    <>
+    
+        <NavBar />
+        <CardRender/>
+    
+    </>
+  )
+}
