+++ src/index.css
@@ -7,7 +7,7 @@
   color: rgba(255, 255, 255, 0.87);
   /* background-color: #242424; */
   /* background-color: #f3f3e7; */
-  background-image: url("./assets/PokePICS/EnterBg.jpg");
+  /* background-image: url("./assets/PokePICS/EnterBg.jpg"); */
   
   font-synthesis: none;
   text-rendering: optimizeLegibility;
